Security hardening: key data copy now ask for permission and has a runner to clear it from memory after some time.
For Keystore destructive operations no messages from remote senders
are allowed or dropped (with exceptions).

Keyring creation and deletion checks if the name is not Master.

Exported keys files now are set with read-write permissions only
for its creator.

Initial support for safe zero-memory for memory that have stored
passwords (in progress).